added extra script enabling either data gerneration from root or conversion into h5
